+++ zenml/backends/orchestrator/aws/orchestrator_aws_backend.py
@@ -17,10 +17,10 @@ import os
 import time
 from typing import Text, Dict, Any
 
-from zenml.core.backends.orchestrator.aws import utils
-from zenml.core.backends.orchestrator import OrchestratorBaseBackend
-from zenml.core.repo import Repository
-from zenml.core.standards import standard_keys as keys
+from zenml.backends.orchestrator.aws import utils
+from zenml.backends.orchestrator import OrchestratorBaseBackend
+from zenml.repo import Repository
+from zenml.standards import standard_keys as keys
 from zenml.utils import path_utils
 from zenml.utils.constants import ZENML_BASE_IMAGE_NAME
 from zenml.utils.logger import get_logger
